Switch to using the docker.elastic.co registry (#582)
The dockerImage property was specified with no registry prefix,
so was implicitly using the default docker.io one. This leads to
issues when trying to pull in newer images such as 7.10.0, which
are present in the docker.elastic.co registry, but not in the
docker.io one.

This fix switches the dockerImage value to the explicit image name
with the registry prefix, as well as adds a compatibility shim. If
the ambiguous image name of "elasticsearch:7.10.0" or a bare tag
is specified, the tag part is used with the default image of
"docker.elastic.co/elasticsearch/elasticsearch".