Upgrade to Keycloak 12.0.4 (#628)
The following things have been done:
- updated to Keycloak 12.0.4
- removed fixed version from test realm file
- introduced property to configure authBasePath
- cleaned up default value handling a bit
- default container command now binds to all interfaces
- changed the wait strategy to listen for port

Also the multiplier of the startup time has been removed as its
intended purpose is not clear and is totally non transparent
to the end user. It initially looked like a bug that the configured
startup timeout did not match before noticing it was actually
doubled all the time.