Enable Non-default images for neo4j (#673)
With this change, a user can use a custom image (for example in my case, one with apoc already installed) or even just one that is forced to be housed on a different server. Obviously, this might not always work with all custom images, but I think the users should be able to test with what they are actually going to be using in production.